src/python3-scrapy/pyproject.toml
[project]
name = "lit-repeat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "scrapy-playwright>=0.0.46",
]
